requirements.txt
chromadb
spacy
pandas
dotenv
requests
flask
flask_sqlalchemy